Add function to create weather info divs for specific times
This commit adds a new function, addWeatherInfoDiv, which dynamically creates weather info divs for specific times. The function takes three parameters: time, icon, and temperature, and constructs a weather info div containing the provided information.

The addWeatherInfoDiv function creates a section element with the class 'weather-info-section' and a div element with the class 'weather-info-div'. It then creates p elements for the time and temperature, and an img element for the weather icon. These elements are appended to the div element, which is then appended to the specified container in the DOM.

This function lays the groundwork for dynamically adding weather info divs for each instance of time in the current day. Later, another function will be implemented to loop through all the times and call addWeatherInfoDiv for each time to display the weather information accordingly.